+++ ScdMergeWizard/Pages/Page10_SourceConnection.Designer.cs
@@ -49,17 +49,18 @@
             // pictureBox1
             // 
             this.pictureBox1.Image = global::ScdMergeWizard.Properties.Resources.database_arrow_right_icon;
-            this.pictureBox1.Margin = new System.Windows.Forms.Padding(4, 4, 4, 4);
-            this.pictureBox1.Size = new System.Drawing.Size(48, 48);
+            this.pictureBox1.Margin = new System.Windows.Forms.Padding(5, 5, 5, 5);
+            this.pictureBox1.Size = new System.Drawing.Size(64, 59);
             // 
             // rtbSrcConnStr
             // 
             this.rtbSrcConnStr.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Left) 
             | System.Windows.Forms.AnchorStyles.Right)));
-            this.rtbSrcConnStr.Location = new System.Drawing.Point(23, 96);
+            this.rtbSrcConnStr.Location = new System.Drawing.Point(31, 118);
+            this.rtbSrcConnStr.Margin = new System.Windows.Forms.Padding(4, 4, 4, 4);
             this.rtbSrcConnStr.Multiline = true;
             this.rtbSrcConnStr.Name = "rtbSrcConnStr";
-            this.rtbSrcConnStr.Size = new System.Drawing.Size(450, 53);
+            this.rtbSrcConnStr.Size = new System.Drawing.Size(599, 64);
             this.rtbSrcConnStr.TabIndex = 9;
             this.rtbSrcConnStr.TextChanged += new System.EventHandler(this.rtbSrcConnStr_TextChanged);
             this.rtbSrcConnStr.Leave += new System.EventHandler(this.rtbSrcConnStr_Leave);
@@ -67,9 +68,10 @@
             // label1
             // 
             this.label1.AutoSize = true;
-            this.label1.Location = new System.Drawing.Point(20, 80);
+            this.label1.Location = new System.Drawing.Point(27, 98);
+            this.label1.Margin = new System.Windows.Forms.Padding(4, 0, 4, 0);
             this.label1.Name = "label1";
-            this.label1.Size = new System.Drawing.Size(134, 13);
+            this.label1.Size = new System.Drawing.Size(163, 16);
             this.label1.TabIndex = 8;
             this.label1.Text = "OLE DB Connection string:";
             // 
@@ -79,9 +81,10 @@
             | System.Windows.Forms.AnchorStyles.Left) 
             | System.Windows.Forms.AnchorStyles.Right)));
             this.rtbCommandText.Enabled = false;
-            this.rtbCommandText.Location = new System.Drawing.Point(23, 228);
+            this.rtbCommandText.Location = new System.Drawing.Point(31, 281);
+            this.rtbCommandText.Margin = new System.Windows.Forms.Padding(4, 4, 4, 4);
             this.rtbCommandText.Name = "rtbCommandText";
-            this.rtbCommandText.Size = new System.Drawing.Size(531, 80);
+            this.rtbCommandText.Size = new System.Drawing.Size(707, 98);
             this.rtbCommandText.TabIndex = 12;
             this.rtbCommandText.Text = "";
             this.rtbCommandText.TextChanged += new System.EventHandler(this.rtbCommandText_TextChanged);
@@ -89,9 +92,10 @@
             // rbIsCommandText
             // 
             this.rbIsCommandText.AutoSize = true;
-            this.rbIsCommandText.Location = new System.Drawing.Point(23, 205);
+            this.rbIsCommandText.Location = new System.Drawing.Point(31, 252);
+            this.rbIsCommandText.Margin = new System.Windows.Forms.Padding(4, 4, 4, 4);
             this.rbIsCommandText.Name = "rbIsCommandText";
-            this.rbIsCommandText.Size = new System.Drawing.Size(95, 17);
+            this.rbIsCommandText.Size = new System.Drawing.Size(113, 20);
             this.rbIsCommandText.TabIndex = 11;
             this.rbIsCommandText.Text = "Command text:";
             this.rbIsCommandText.UseVisualStyleBackColor = true;
@@ -101,9 +105,10 @@
             // 
             this.rbIsTableOrView.AutoSize = true;
             this.rbIsTableOrView.Checked = true;
-            this.rbIsTableOrView.Location = new System.Drawing.Point(23, 155);
+            this.rbIsTableOrView.Location = new System.Drawing.Point(31, 191);
+            this.rbIsTableOrView.Margin = new System.Windows.Forms.Padding(4, 4, 4, 4);
             this.rbIsTableOrView.Name = "rbIsTableOrView";
-            this.rbIsTableOrView.Size = new System.Drawing.Size(142, 17);
+            this.rbIsTableOrView.Size = new System.Drawing.Size(173, 20);
             this.rbIsTableOrView.TabIndex = 10;
             this.rbIsTableOrView.TabStop = true;
             this.rbIsTableOrView.Text = "Table, View or Synonym:";
@@ -117,9 +122,10 @@
             this.cbxSrcTableOrViewName.DrawMode = System.Windows.Forms.DrawMode.OwnerDrawVariable;
             this.cbxSrcTableOrViewName.DropDownStyle = System.Windows.Forms.ComboBoxStyle.DropDownList;
             this.cbxSrcTableOrViewName.FormattingEnabled = true;
-            this.cbxSrcTableOrViewName.Location = new System.Drawing.Point(23, 178);
+            this.cbxSrcTableOrViewName.Location = new System.Drawing.Point(31, 219);
+            this.cbxSrcTableOrViewName.Margin = new System.Windows.Forms.Padding(4, 4, 4, 4);
             this.cbxSrcTableOrViewName.Name = "cbxSrcTableOrViewName";
-            this.cbxSrcTableOrViewName.Size = new System.Drawing.Size(416, 21);
+            this.cbxSrcTableOrViewName.Size = new System.Drawing.Size(553, 23);
             this.cbxSrcTableOrViewName.TabIndex = 13;
             this.cbxSrcTableOrViewName.DropDown += new System.EventHandler(this.cbxSrcTables_DropDown);
             this.cbxSrcTableOrViewName.SelectedIndexChanged += new System.EventHandler(this.cbxSrcTableOrViewName_SelectedIndexChanged);
@@ -128,9 +134,10 @@
             // buttonPreviewData
             // 
             this.buttonPreviewData.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Right)));
-            this.buttonPreviewData.Location = new System.Drawing.Point(479, 176);
+            this.buttonPreviewData.Location = new System.Drawing.Point(639, 217);
+            this.buttonPreviewData.Margin = new System.Windows.Forms.Padding(4, 4, 4, 4);
             this.buttonPreviewData.Name = "buttonPreviewData";
-            this.buttonPreviewData.Size = new System.Drawing.Size(75, 23);
+            this.buttonPreviewData.Size = new System.Drawing.Size(100, 28);
             this.buttonPreviewData.TabIndex = 14;
             this.buttonPreviewData.Text = "Preview...";
             this.buttonPreviewData.UseVisualStyleBackColor = true;
@@ -140,9 +147,10 @@
             // 
             this.buttonFilter.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Right)));
             this.buttonFilter.Image = global::ScdMergeWizard.Properties.Resources.Filter2HS;
-            this.buttonFilter.Location = new System.Drawing.Point(445, 176);
+            this.buttonFilter.Location = new System.Drawing.Point(593, 217);
+            this.buttonFilter.Margin = new System.Windows.Forms.Padding(4, 4, 4, 4);
             this.buttonFilter.Name = "buttonFilter";
-            this.buttonFilter.Size = new System.Drawing.Size(28, 23);
+            this.buttonFilter.Size = new System.Drawing.Size(37, 28);
             this.buttonFilter.TabIndex = 16;
             this.buttonFilter.UseVisualStyleBackColor = true;
             this.buttonFilter.Click += new System.EventHandler(this.buttonFilter_Click);
@@ -152,9 +160,10 @@
             this.labelFilter.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Left) 
             | System.Windows.Forms.AnchorStyles.Right)));
             this.labelFilter.Font = new System.Drawing.Font("Microsoft Sans Serif", 8.25F, System.Drawing.FontStyle.Italic, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
-            this.labelFilter.Location = new System.Drawing.Point(184, 157);
+            this.labelFilter.Location = new System.Drawing.Point(245, 193);
+            this.labelFilter.Margin = new System.Windows.Forms.Padding(4, 0, 4, 0);
             this.labelFilter.Name = "labelFilter";
-            this.labelFilter.Size = new System.Drawing.Size(255, 15);
+            this.labelFilter.Size = new System.Drawing.Size(340, 18);
             this.labelFilter.TabIndex = 17;
             this.labelFilter.Text = "Filter: (none)";
             this.labelFilter.TextAlign = System.Drawing.ContentAlignment.MiddleRight;
@@ -165,10 +174,10 @@
             this.btnEditCnn.ContextMenuStrip = this.cmsEdit;
             this.btnEditCnn.ImageAlign = System.Drawing.ContentAlignment.MiddleRight;
             this.btnEditCnn.ImageKey = "Normal";
-            this.btnEditCnn.Location = new System.Drawing.Point(480, 96);
-            this.btnEditCnn.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
+            this.btnEditCnn.Location = new System.Drawing.Point(640, 118);
+            this.btnEditCnn.Margin = new System.Windows.Forms.Padding(3, 2, 3, 2);
             this.btnEditCnn.Name = "btnEditCnn";
-            this.btnEditCnn.Size = new System.Drawing.Size(74, 28);
+            this.btnEditCnn.Size = new System.Drawing.Size(99, 34);
             this.btnEditCnn.TabIndex = 18;
             this.btnEditCnn.Text = "Edit";
             this.btnEditCnn.UseVisualStyleBackColor = true;
@@ -183,7 +192,7 @@
             this.cmsEdit.RightToLeft = System.Windows.Forms.RightToLeft.Yes;
             this.cmsEdit.ShowCheckMargin = true;
             this.cmsEdit.ShowImageMargin = false;
-            this.cmsEdit.Size = new System.Drawing.Size(267, 48);
+            this.cmsEdit.Size = new System.Drawing.Size(267, 70);
             // 
             // tsmiEditNew
             // 
@@ -199,11 +208,12 @@
             this.tsmiEditOld.Name = "tsmiEditOld";
             this.tsmiEditOld.Size = new System.Drawing.Size(266, 22);
             this.tsmiEditOld.Text = "Edit (OLE DB Connection)";
+            this.tsmiEditOld.Visible = false;
             this.tsmiEditOld.Click += new System.EventHandler(this.tsmiEditOld_Click);
             // 
             // Page10_SourceConnection
             // 
-            this.AutoScaleDimensions = new System.Drawing.SizeF(6F, 13F);
+            this.AutoScaleDimensions = new System.Drawing.SizeF(8F, 16F);
             this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
             this.Controls.Add(this.btnEditCnn);
             this.Controls.Add(this.labelFilter);
@@ -216,9 +226,9 @@
             this.Controls.Add(this.rbIsCommandText);
             this.Controls.Add(this.rbIsTableOrView);
             this.Description = "Define the source of your data";
-            this.Margin = new System.Windows.Forms.Padding(4, 4, 4, 4);
+            this.Margin = new System.Windows.Forms.Padding(5, 5, 5, 5);
             this.Name = "Page10_SourceConnection";
-            this.Size = new System.Drawing.Size(588, 339);
+            this.Size = new System.Drawing.Size(784, 417);
             this.Title = "Source Connection";
             this.Controls.SetChildIndex(this.rbIsTableOrView, 0);
             this.Controls.SetChildIndex(this.rbIsCommandText, 0);