+++ ScdMergeWizard/Properties/AssemblyInfo.cs
@@ -6,11 +6,11 @@ using System.Runtime.InteropServices;
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
 [assembly: AssemblyTitle("ScdMergeWizard")]
-[assembly: AssemblyDescription("Slowly Changing Dimension (SCD) Merge Wizard is used to generate and maintain T-SQL (MERGE) statement for SCD operations. \nBy Miljan Radovic (knyazs) http://si.linkedin.com/in/miljanradovic \nand Kamil Nowinski http://SQLPlayer.net\n\nFor further information, please visit:\nhttps://github.com/SQLPlayer/SCD-Merge-Wizard")]
+[assembly: AssemblyDescription("Slowly Changing Dimension (SCD) Merge Wizard is used to generate and maintain T-SQL (MERGE) statement for SCD operations. \nBy Miljan Radovic (knyazs) http://si.linkedin.com/in/miljanradovic \nand Kamil Nowinski http://AzurePlayer.net\n\nFor further information, please visit:\nhttps://github.com/Azure-Player/SCD-Merge-Wizard")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
 [assembly: AssemblyProduct("ScdMergeWizard")]
-[assembly: AssemblyCopyright("Copyright © 2013-2021")]
+[assembly: AssemblyCopyright("Copyright © 2013-2024")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 
@@ -32,5 +32,5 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers 
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("4.3.1.0")]
-[assembly: AssemblyFileVersion("4.3.1.0")]
+[assembly: AssemblyVersion("5.0.0.0")]
+[assembly: AssemblyFileVersion("5.0.0.0")]