+++ Release-Notes.md
@@ -1,3 +1,9 @@
+## [5.0.0] - 2024-09-04
+- Updated to use .NET Framework 4.8 (Windows 10/11)
+- Removed old connection method window
+- Replaced `System.Data.SqlClient` with `Microsoft.Data.SqlClient`
+- Added new authentication method: **Entra with MFA**
+
 ## [4.3.0] - 2020-08-06
 * Added space before OR/AND in a condition statement (#9)
 * Change: List of tables is sorted now (#2)